Add custom dump
