+++ Example/GoTrue Example/Helpers/Stringfy.swift
@@ -1,0 +1,8 @@
+import CustomDump
+import Foundation
+
+func stringfy(_ value: Any) -> String {
+  var output = ""
+  customDump(value, to: &output)
+  return output
+}
