+++ Example/GoTrue Example/Sign up with email and password/SignUpWithEmailAndPasswordView.swift
@@ -1,27 +1,6 @@
 import GoTrueHTTP
 import SwiftUI
 
-enum ActionStatus<Success, Failure: Error> {
-  case idle, inFlight
-  case success(Success)
-  case failure(Failure)
-
-  var isInFlight: Bool {
-    if case .inFlight = self { return true }
-    return false
-  }
-
-  var success: Success? {
-    if case .success(let value) = self { return value }
-    return nil
-  }
-
-  var failure: Failure? {
-    if case .failure(let error) = self { return error }
-    return nil
-  }
-}
-
 struct SignUpWithEmailAndPasswordView: View {
   static let title = "Sign up with email and password"
 
@@ -67,7 +46,7 @@ struct SignUpWithEmailAndPasswordView: View {
         }
 
         if let error = status.failure {
-          Text(error.localizedDescription)
+          Text(stringfy(error))
             .font(.footnote)
             .foregroundColor(.red)
         }
@@ -86,21 +65,6 @@ struct SignUpWithEmailAndPasswordView: View {
         status = .failure(error)
       }
     }
-  }
-
-  private func stringfy(_ response: Paths.Signup.PostResponse) -> String {
-    let encoder = JSONEncoder()
-    encoder.outputFormatting = [.prettyPrinted, .sortedKeys]
-    var data: Data?
-
-    switch response {
-    case .session(let session):
-      data = try? encoder.encode(session)
-    case .user(let user):
-      data = try? encoder.encode(user)
-    }
-
-    return data.flatMap { String(data: $0, encoding: .utf8) } ?? ""
   }
 }
